refactor(cosmo-pd101): extract useFxModuleController hook, deduplicate custom FX renderers
Extract shared FX module controller logic (selectedPreset, params, enabled,
handlePresetChange, store bindings) into useFxModuleController hook. All 4
custom FX renderers (Delay, PhaseMod, Tremolo, Vibrato) delegate to it,
removing ~25 lines of identical boilerplate per renderer.
