+++ packages/cosmo-pd101/src/components/panels/drawer-modules/custom/VibratoModuleRenderer.tsx
@@ -1,4 +1,3 @@
-import { useState } from "react";
 import Button from "@/components/controls/Button";
 import ControlKnob from "@/components/controls/ControlKnob";
 import {
@@ -7,14 +6,11 @@ import {
 	getKnobControl,
 	getModDestinationByParam,
 	getTooltip,
-	resolveEnabled,
-	resolvePresetPatchParams,
 } from "@/components/panels/drawer-modules/custom/utils";
+import { useFxModuleController } from "@/components/panels/drawer-modules/custom/useFxModuleController";
 import type { FxSlotModuleConfig } from "@/components/panels/drawer-modules/fxSlotModuleConfig";
 import ModuleFrame from "@/components/primitives/ModuleFrame";
 import ModulePresetPopover from "@/components/primitives/ModulePresetPopover";
-import { requestApplyModulePreset } from "@/features/synth/engine/modulePresetEvents";
-import { useSynthStore } from "@/features/synth/synthStore";
 
 export default function VibratoModuleRenderer({
 	config,
@@ -23,45 +19,20 @@ export default function VibratoModuleRenderer({
 	config: FxSlotModuleConfig;
 	slot: number;
 }) {
-	const [selectedPreset, setSelectedPreset] = useState<string>("");
-	const rawSlot = useSynthStore((state) => state.fxSlots[slot]);
-	const setFxSlotParams = useSynthStore((state) => state.setFxSlotParams);
+	const {
+		selectedPreset,
+		setFxSlotParams,
+		params,
+		enabled,
+		handlePresetChange,
+	} = useFxModuleController(config, slot);
 
-	if (rawSlot?.type !== config.type) {
-		return null;
-	}
-
-	const params = (rawSlot as { params: Record<string, unknown> }).params;
-	const enabled = resolveEnabled(params);
 	const waveformControl = getButtonGroupControl(config, "waveform");
 	const rateControl = getKnobControl(config, "rate");
 	const depthControl = getKnobControl(config, "depth");
 	const delayControl = getKnobControl(config, "delay");
 	const waveformValue = asNumber(params.waveform, 1);
 	const modDestinationByParam = getModDestinationByParam(config.type);
-
-	const handlePresetChange = (presetId: string) => {
-		setSelectedPreset(presetId);
-		const preset = config.presets.find((entry) => entry.id === presetId);
-		if (!preset) {
-			return;
-		}
-
-		const patchParams = resolvePresetPatchParams(
-			config,
-			preset.patch as Record<string, unknown>,
-		);
-		if (!patchParams) {
-			return;
-		}
-
-		setFxSlotParams(slot, patchParams);
-		requestApplyModulePreset({
-			module: config.moduleKey,
-			preset: preset.id,
-			patch: preset.patch,
-		});
-	};
 
 	return (
 		<ModuleFrame
